requirements.txt
streamlit>=1.32,<2.0
pyairtable>=2.2,<3.0
qrcode[pil]>=7.4,<8.0
Pillow>=10.0,<12.0
pyzbar>=0.1.9,<0.2  # optional, only if you want automatic decode on the server (requires zbar on the system)
